requirements.txt
tensorflow-cpu==2.6.0
torch==1.12.0
fairseq==0.9.0
numpy
